added legend for move category and refactored code
