+++ README.md
@@ -11,8 +11,8 @@ Inspired by [David](https://github.com/davidhckh)
 ## TODO
 
 - add page to save pokemon in (team)
-- add learnable moves in pokemon details page
 - fix scroll down indication (see eevee)
 - add is_legendary tag
 - organise response types into folders similar to [this package](https://www.npmjs.com/package/@bgoff1/pokeapi-types?activeTab=code).
 - add a loader to pokemon moves
+- extract popper code into a component
