+++ src/pages/pokemon-details-page/moves/moves-table/types.ts
@@ -1,0 +1,8 @@
+import { LevelUpRowInfoType } from "../types";
+
+export type Order = "asc" | "desc";
+
+export type MovesTableType = {
+  data: LevelUpRowInfoType[];
+  method: string;
+};
