+++ src/utils/helpers.ts
@@ -77,3 +77,46 @@ export const getDataPromises = <T>(
     })
   );
 };
+
+/**
+ * for sorting an array in descending order (used for MOVES TABLE)
+ */
+const descendingComparator = <T extends any>(
+  a: T,
+  b: T,
+  orderBy: keyof T
+): number => {
+  if (b[orderBy] === null) return -1;
+  if (a[orderBy] === null) return 1;
+  if (b[orderBy] < a[orderBy]) {
+    return -1;
+  }
+  if (b[orderBy] > a[orderBy]) {
+    return 1;
+  }
+  return 0;
+};
+
+export const getComparator = <T extends any, K extends any>(
+  order: K,
+  orderBy: keyof T
+): ((a: T, b: T) => number) => {
+  return order === "desc"
+    ? (a, b) => descendingComparator(a, b, orderBy)
+    : (a, b) => -descendingComparator(a, b, orderBy);
+};
+
+export const stableSort = <T extends any>(
+  array: T[],
+  comparator: (a: T, b: T) => number
+) => {
+  const stabilizedThis = array.map((el, index) => [el, index] as [T, number]);
+  stabilizedThis.sort((a, b) => {
+    const order = comparator(a[0], b[0]);
+    if (order !== 0) {
+      return order;
+    }
+    return a[1] - b[1];
+  });
+  return stabilizedThis.map((el) => el[0]);
+};
