+++ src/utils/theme.ts
@@ -1,0 +1,41 @@
+import { createTheme } from "@mui/material";
+
+export const theme = createTheme({
+  shadows: [
+    "none",
+    "5px 5px #e8ebf1",
+    "5px 5px #e0e3e9",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+    "",
+  ],
+  palette: {
+    primary: {
+      light: "#f0f3f8",
+      main: "#c3c9cf",
+      400: "#e0e3e9",
+    },
+    text: {
+      primary: "#0c214a",
+    },
+  },
+});
